style: Make images smaller and centered
With this commit, images (specifically the one for How Kelda Works) are limited
to 70% of the block width on big screens. This prevents them from being
overwhelmingly big, even if the source image is large.
Now that images don't take up the full horizontal space, we make sure to center
them.